Add files via upload
Added two files. One is a csv of student data. The other is a python file that reads the csv and analyzes the student data. Specifically, whether or not parental occupation as a teacher will affect the grades of their children as students, and if the level of education of parents influences the desire to pursue higher education for children.